refactor singleton, remove interfaces on entities, refactor builder

+++ src/com/excilys/computerdatabase/controller/DetailsComputerController.java
@@ -52,7 +52,7 @@ public class DetailsComputerController {
 	/**
 	 * @param pOperation
 	 */
-	public void makeOperation(String pOperation, int pId) {
+	public void makeOperation(String pOperation, long pId) {
 		switch (pOperation)
 		{
 			case "d":
